+++ changelog.md
@@ -1,3 +1,9 @@
+### 0.10.0 (not yet released)
+
+- bump bioimageio.spec library version to 0.5.8.0
+- support Keras v3 weights descriptions
+- new `bioimageio predict-block-artifacts` command
+
 ### 0.9.6
 
 - bump bioimageio.spec library version to 0.5.7.4